@@ -7,6 +7,9 @@
     <groupId>com.yoursole1</groupId>
     <artifactId>moderation-bot</artifactId>
     <version>1.0-SNAPSHOT</version>
+    <properties>
+        <slf4j.version>1.7.32</slf4j.version>
+    </properties>
 
     <build>
         <plugins>
@@ -18,7 +21,7 @@
                         <manifest>
                             <addClasspath>true</addClasspath>
                             <classpathPrefix>lib/</classpathPrefix>
-                            <mainClass>main.main</mainClass>
+                            <mainClass>com.yoursole1.Main</mainClass>
                         </manifest>
                     </archive>
                 </configuration>
@@ -36,7 +39,7 @@
                         <configuration>
                             <transformers>
                                 <transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                                    <mainClass>main.main</mainClass>
+                                    <mainClass>com.yoursole1.Main</mainClass>
                                 </transformer>
                             </transformers>
                         </configuration>
@@ -47,8 +50,8 @@
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-compiler-plugin</artifactId>
                 <configuration>
-                    <source>9</source>
-                    <target>9</target>
+                    <source>14</source>
+                    <target>14</target>
                 </configuration>
             </plugin>
         </plugins>
@@ -67,12 +70,22 @@
         <dependency>
             <groupId>com.google.code.gson</groupId>
             <artifactId>gson</artifactId>
-            <version>2.3</version>
+            <version>2.8.7</version>
         </dependency>
         <dependency>
             <groupId>org.projectlombok</groupId>
             <artifactId>lombok</artifactId>
             <version>1.18.20</version>
+        </dependency>
+        <dependency>
+            <groupId>org.slf4j</groupId>
+            <artifactId>slf4j-api</artifactId>
+            <version>${slf4j.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.slf4j</groupId>
+            <artifactId>slf4j-simple</artifactId>
+            <version>${slf4j.version}</version>
         </dependency>
     </dependencies>
     <repositories>
