@@ -3,15 +3,19 @@ package com.yoursole1.command;
 import com.yoursole1.command.annotation.AnnotationProcessor;
 import com.yoursole1.command.exception.CommandException;
 import com.yoursole1.command.exception.CommandNotFoundException;
+import com.yoursole1.command.exception.UnableToInvokeCommandException;
+import lombok.extern.slf4j.Slf4j;
 import net.dv8tion.jda.api.events.message.guild.GuildMessageReceivedEvent;
 
 import java.lang.reflect.InvocationTargetException;
 import java.util.List;
+import java.util.Objects;
 
 /**
  * @author Owen Hayes
  */
-public class CommandParser {
+@Slf4j
+public final class CommandParser {
 
     private static List<CommandWrapper> commands;
 
@@ -19,38 +23,48 @@ public class CommandParser {
         try {
             commands = AnnotationProcessor.loadCommands(CommandRegistry.class);
         } catch (CommandException e) {
-            e.printStackTrace();
+            log.error("Commands could not be loaded", e);
         }
     }
+
+    private CommandParser() {}
 
     public static List<CommandWrapper> getCommands() {
         return List.copyOf(commands);
     }
 
-    public static CommandWrapper findCommand(String message) throws CommandNotFoundException {
+    public static CommandWrapper findCommand(String message) {
         for (CommandWrapper command : commands) {
             if (command.commandStringMatches(message)) {
                 return command;
             }
         }
-        throw new CommandNotFoundException();
+        return null;
     }
 
     public static CommandWrapper findCommand(GuildMessageReceivedEvent event) throws CommandNotFoundException {
         return findCommand(event.getMessage().getContentRaw());
     }
 
-    public static String parse(String input) throws
-            InvocationTargetException,
-            IllegalAccessException {
+    public static String parse(String input) throws UnableToInvokeCommandException {
         CommandWrapper command = findCommand(input);
-        return command.execute(input);
+        if (Objects.isNull(command)) {
+            return null;
+        } else {
+            return command.execute(input);
+        }
     }
 
-    public static String parse(GuildMessageReceivedEvent event) throws
-            InvocationTargetException,
-            IllegalAccessException {
-        return parse(event.getMessage().getContentRaw());
+    public static String parse(GuildMessageReceivedEvent event) throws UnableToInvokeCommandException {
+        String input = event.getMessage().getContentRaw();
+        CommandWrapper command = findCommand(input);
+        if (Objects.isNull(command)) {
+            return null;
+        } else if (!command.checkPermissions(event)) {
+            return "Insufficient permissions";
+        } else {
+            return command.execute(input);
+        }
     }
 
 
