@@ -1,15 +1,18 @@
 package com.yoursole1.command.annotation;
 
-import java.lang.annotation.*;
+import java.lang.annotation.Documented;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
 
 @Documented
 @Target(ElementType.METHOD)
 @Retention(RetentionPolicy.RUNTIME)
 public @interface Command {
-    long MAX_PERMISSIONS = 2146958591;
 
     String prefix() default "!";
     String exampleText() default "";
     String helpText() default "";
-    long permissions() default MAX_PERMISSIONS;
+    String permissions() default "admin";
 }
