@@ -1,6 +1,11 @@
 package com.yoursole1.command.annotation;
 
-import java.lang.annotation.*;
+import java.lang.annotation.Documented;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
 
 @Documented
 @Target(ElementType.TYPE)
