@@ -1,4 +1,4 @@
-package main.repository;
+package com.yoursole1.repository;
 
 public interface ModeratorRepository extends StaticFileRepository {
 }
